Make config compatible with github actions
Changes various checks (apptainer tmpdir and correspondance between
actual current partition and requested partition profile) so that they
are skipped when the host machine is not the VSC CalcUA.

This was done so that the github actions no longer fail. Specifically,
it stops the `nextflow config -show-profiles` command from quitting out
on the `System.exit(1)` statements.